Update and rename Task1.java to Tasks1to5.java
